# Filter out read-only instances, but skip filtering when it's not required.