… ring/ring_test.go