**What this PR does**:

This PR updates shuffle-sharding to 1) filter out read-only instances on shuffle shard with lookback, if instance changed its state before lookback window, 2) avoid computing shuffle shard if no filtering is required.

Benchmark compared to `main` (updated after https://github.com/grafana/dskit/pull/565/commits/95e95648d2c85553387565584c8e394ec79eaf0f):

```
benchstat main-new-bench.txt final2.txt
goos: darwin
goarch: arm64
pkg: github.com/grafana/dskit/ring
                                                                                  │ main-new-bench.txt │              final2.txt              │
                                                                                  │       sec/op       │    sec/op     vs base                │
Ring_ShuffleShard/num_instances_=_50,_num_zones_=_1,_shard_size_=_0-10                657443.00n ± ∞ ¹   59.71n ± ∞ ¹   -99.99% (p=0.008 n=5)
Ring_ShuffleShard/num_instances_=_50,_num_zones_=_1,_shard_size_=_3-10                    19.83µ ± ∞ ¹   19.34µ ± ∞ ¹         ~ (p=0.151 n=5)
Ring_ShuffleShard/num_instances_=_50,_num_zones_=_1,_shard_size_=_10-10                   53.21µ ± ∞ ¹   53.59µ ± ∞ ¹         ~ (p=0.841 n=5)
Ring_ShuffleShard/num_instances_=_50,_num_zones_=_1,_shard_size_=_30-10                   163.3µ ± ∞ ¹   163.5µ ± ∞ ¹         ~ (p=0.310 n=5)
Ring_ShuffleShard/num_instances_=_50,_num_zones_=_1,_shard_size_=_100-10                  651.8µ ± ∞ ¹   272.4µ ± ∞ ¹   -58.21% (p=0.008 n=5)
Ring_ShuffleShard/num_instances_=_50,_num_zones_=_1,_shard_size_=_1000-10                 652.2µ ± ∞ ¹   271.9µ ± ∞ ¹   -58.31% (p=0.008 n=5)
Ring_ShuffleShard/num_instances_=_50,_num_zones_=_3,_shard_size_=_0-10                680198.00n ± ∞ ¹   58.80n ± ∞ ¹   -99.99% (p=0.008 n=5)
Ring_ShuffleShard/num_instances_=_50,_num_zones_=_3,_shard_size_=_3-10                    45.55µ ± ∞ ¹   45.19µ ± ∞ ¹    -0.78% (p=0.008 n=5)
Ring_ShuffleShard/num_instances_=_50,_num_zones_=_3,_shard_size_=_10-10                   87.33µ ± ∞ ¹   89.49µ ± ∞ ¹    +2.47% (p=0.016 n=5)
Ring_ShuffleShard/num_instances_=_50,_num_zones_=_3,_shard_size_=_30-10                   200.8µ ± ∞ ¹   203.2µ ± ∞ ¹         ~ (p=0.222 n=5)
Ring_ShuffleShard/num_instances_=_50,_num_zones_=_3,_shard_size_=_100-10                  677.0µ ± ∞ ¹   293.9µ ± ∞ ¹   -56.59% (p=0.008 n=5)
Ring_ShuffleShard/num_instances_=_50,_num_zones_=_3,_shard_size_=_1000-10                 671.4µ ± ∞ ¹   291.4µ ± ∞ ¹   -56.60% (p=0.008 n=5)
Ring_ShuffleShard/num_instances_=_100,_num_zones_=_1,_shard_size_=_0-10              1395995.00n ± ∞ ¹   59.98n ± ∞ ¹  -100.00% (p=0.008 n=5)
Ring_ShuffleShard/num_instances_=_100,_num_zones_=_1,_shard_size_=_3-10                   19.45µ ± ∞ ¹   19.35µ ± ∞ ¹         ~ (p=0.310 n=5)
Ring_ShuffleShard/num_instances_=_100,_num_zones_=_1,_shard_size_=_10-10                  53.32µ ± ∞ ¹   55.47µ ± ∞ ¹    +4.02% (p=0.008 n=5)
Ring_ShuffleShard/num_instances_=_100,_num_zones_=_1,_shard_size_=_30-10                  162.9µ ± ∞ ¹   168.1µ ± ∞ ¹    +3.19% (p=0.008 n=5)
Ring_ShuffleShard/num_instances_=_100,_num_zones_=_1,_shard_size_=_100-10                 654.3µ ± ∞ ¹   635.1µ ± ∞ ¹    -2.93% (p=0.008 n=5)
Ring_ShuffleShard/num_instances_=_100,_num_zones_=_1,_shard_size_=_1000-10               1366.3µ ± ∞ ¹   628.9µ ± ∞ ¹   -53.97% (p=0.008 n=5)
Ring_ShuffleShard/num_instances_=_100,_num_zones_=_3,_shard_size_=_0-10              1377139.00n ± ∞ ¹   60.41n ± ∞ ¹  -100.00% (p=0.008 n=5)
Ring_ShuffleShard/num_instances_=_100,_num_zones_=_3,_shard_size_=_3-10                   45.29µ ± ∞ ¹   46.29µ ± ∞ ¹    +2.22% (p=0.008 n=5)
Ring_ShuffleShard/num_instances_=_100,_num_zones_=_3,_shard_size_=_10-10                  86.79µ ± ∞ ¹   91.29µ ± ∞ ¹    +5.18% (p=0.008 n=5)
Ring_ShuffleShard/num_instances_=_100,_num_zones_=_3,_shard_size_=_30-10                  200.1µ ± ∞ ¹   208.0µ ± ∞ ¹    +3.94% (p=0.008 n=5)
Ring_ShuffleShard/num_instances_=_100,_num_zones_=_3,_shard_size_=_100-10                1147.1µ ± ∞ ¹   672.6µ ± ∞ ¹   -41.36% (p=0.008 n=5)
Ring_ShuffleShard/num_instances_=_100,_num_zones_=_3,_shard_size_=_1000-10               1383.6µ ± ∞ ¹   671.2µ ± ∞ ¹   -51.49% (p=0.008 n=5)
Ring_ShuffleShard/num_instances_=_1000,_num_zones_=_1,_shard_size_=_0-10            17568050.00n ± ∞ ¹   61.36n ± ∞ ¹  -100.00% (p=0.008 n=5)
Ring_ShuffleShard/num_instances_=_1000,_num_zones_=_1,_shard_size_=_3-10                  17.58µ ± ∞ ¹   18.28µ ± ∞ ¹    +3.96% (p=0.008 n=5)
Ring_ShuffleShard/num_instances_=_1000,_num_zones_=_1,_shard_size_=_10-10                 50.52µ ± ∞ ¹   52.53µ ± ∞ ¹    +3.99% (p=0.016 n=5)
Ring_ShuffleShard/num_instances_=_1000,_num_zones_=_1,_shard_size_=_30-10                 157.3µ ± ∞ ¹   163.3µ ± ∞ ¹    +3.82% (p=0.008 n=5)
Ring_ShuffleShard/num_instances_=_1000,_num_zones_=_1,_shard_size_=_100-10                626.2µ ± ∞ ¹   651.5µ ± ∞ ¹    +4.04% (p=0.008 n=5)
Ring_ShuffleShard/num_instances_=_1000,_num_zones_=_1,_shard_size_=_1000-10               8.996m ± ∞ ¹   8.596m ± ∞ ¹    -4.44% (p=0.008 n=5)
Ring_ShuffleShard/num_instances_=_1000,_num_zones_=_3,_shard_size_=_0-10            17422654.00n ± ∞ ¹   61.11n ± ∞ ¹  -100.00% (p=0.008 n=5)
Ring_ShuffleShard/num_instances_=_1000,_num_zones_=_3,_shard_size_=_3-10                  41.42µ ± ∞ ¹   43.73µ ± ∞ ¹    +5.59% (p=0.008 n=5)
Ring_ShuffleShard/num_instances_=_1000,_num_zones_=_3,_shard_size_=_10-10                 80.49µ ± ∞ ¹   86.60µ ± ∞ ¹    +7.60% (p=0.008 n=5)
Ring_ShuffleShard/num_instances_=_1000,_num_zones_=_3,_shard_size_=_30-10                 191.4µ ± ∞ ¹   200.6µ ± ∞ ¹    +4.80% (p=0.008 n=5)
Ring_ShuffleShard/num_instances_=_1000,_num_zones_=_3,_shard_size_=_100-10                692.7µ ± ∞ ¹   719.2µ ± ∞ ¹         ~ (p=0.151 n=5)
Ring_ShuffleShard/num_instances_=_1000,_num_zones_=_3,_shard_size_=_1000-10              17.892m ± ∞ ¹   8.968m ± ∞ ¹   -49.87% (p=0.008 n=5)
Ring_ShuffleShardCached/num_instances_=_50,_num_zones_=_1,_shard_size_=_0-10            1688.00n ± ∞ ¹   68.99n ± ∞ ¹   -95.91% (p=0.008 n=5)
Ring_ShuffleShardCached/num_instances_=_50,_num_zones_=_1,_shard_size_=_3-10              149.4n ± ∞ ¹   148.7n ± ∞ ¹         ~ (p=0.802 n=5)
Ring_ShuffleShardCached/num_instances_=_50,_num_zones_=_1,_shard_size_=_10-10             366.3n ± ∞ ¹   357.0n ± ∞ ¹    -2.54% (p=0.032 n=5)
Ring_ShuffleShardCached/num_instances_=_50,_num_zones_=_1,_shard_size_=_30-10             1.027µ ± ∞ ¹   1.008µ ± ∞ ¹         ~ (p=0.151 n=5)
Ring_ShuffleShardCached/num_instances_=_50,_num_zones_=_1,_shard_size_=_100-10            1.671µ ± ∞ ¹   1.693µ ± ∞ ¹         ~ (p=0.548 n=5)
Ring_ShuffleShardCached/num_instances_=_50,_num_zones_=_1,_shard_size_=_1000-10           1.700µ ± ∞ ¹   1.648µ ± ∞ ¹    -3.06% (p=0.008 n=5)
Ring_ShuffleShardCached/num_instances_=_50,_num_zones_=_3,_shard_size_=_0-10            1698.00n ± ∞ ¹   69.67n ± ∞ ¹   -95.90% (p=0.008 n=5)
Ring_ShuffleShardCached/num_instances_=_50,_num_zones_=_3,_shard_size_=_3-10              149.1n ± ∞ ¹   148.5n ± ∞ ¹         ~ (p=0.968 n=5)
Ring_ShuffleShardCached/num_instances_=_50,_num_zones_=_3,_shard_size_=_10-10             431.9n ± ∞ ¹   426.0n ± ∞ ¹         ~ (p=0.310 n=5)
Ring_ShuffleShardCached/num_instances_=_50,_num_zones_=_3,_shard_size_=_30-10             1.015µ ± ∞ ¹   1.009µ ± ∞ ¹         ~ (p=0.103 n=5)
Ring_ShuffleShardCached/num_instances_=_50,_num_zones_=_3,_shard_size_=_100-10            1.691µ ± ∞ ¹   1.648µ ± ∞ ¹         ~ (p=0.151 n=5)
Ring_ShuffleShardCached/num_instances_=_50,_num_zones_=_3,_shard_size_=_1000-10           1.630µ ± ∞ ¹   1.668µ ± ∞ ¹         ~ (p=0.421 n=5)
Ring_ShuffleShardCached/num_instances_=_100,_num_zones_=_1,_shard_size_=_0-10           3602.00n ± ∞ ¹   69.57n ± ∞ ¹   -98.07% (p=0.008 n=5)
Ring_ShuffleShardCached/num_instances_=_100,_num_zones_=_1,_shard_size_=_3-10             147.5n ± ∞ ¹   147.0n ± ∞ ¹         ~ (p=0.690 n=5)
Ring_ShuffleShardCached/num_instances_=_100,_num_zones_=_1,_shard_size_=_10-10            360.3n ± ∞ ¹   365.2n ± ∞ ¹         ~ (p=0.222 n=5)
Ring_ShuffleShardCached/num_instances_=_100,_num_zones_=_1,_shard_size_=_30-10            1.027µ ± ∞ ¹   1.031µ ± ∞ ¹         ~ (p=0.690 n=5)
Ring_ShuffleShardCached/num_instances_=_100,_num_zones_=_1,_shard_size_=_100-10           3.604µ ± ∞ ¹   3.577µ ± ∞ ¹         ~ (p=1.000 n=5)
Ring_ShuffleShardCached/num_instances_=_100,_num_zones_=_1,_shard_size_=_1000-10          3.570µ ± ∞ ¹   3.581µ ± ∞ ¹         ~ (p=1.000 n=5)
Ring_ShuffleShardCached/num_instances_=_100,_num_zones_=_3,_shard_size_=_0-10           3608.00n ± ∞ ¹   69.46n ± ∞ ¹   -98.07% (p=0.008 n=5)
Ring_ShuffleShardCached/num_instances_=_100,_num_zones_=_3,_shard_size_=_3-10             149.9n ± ∞ ¹   150.3n ± ∞ ¹         ~ (p=0.421 n=5)
Ring_ShuffleShardCached/num_instances_=_100,_num_zones_=_3,_shard_size_=_10-10            424.4n ± ∞ ¹   429.2n ± ∞ ¹         ~ (p=0.841 n=5)
Ring_ShuffleShardCached/num_instances_=_100,_num_zones_=_3,_shard_size_=_30-10            1.014µ ± ∞ ¹   1.028µ ± ∞ ¹         ~ (p=1.000 n=5)
Ring_ShuffleShardCached/num_instances_=_100,_num_zones_=_3,_shard_size_=_100-10           3.606µ ± ∞ ¹   3.580µ ± ∞ ¹    -0.72% (p=0.032 n=5)
Ring_ShuffleShardCached/num_instances_=_100,_num_zones_=_3,_shard_size_=_1000-10          3.682µ ± ∞ ¹   3.545µ ± ∞ ¹    -3.72% (p=0.016 n=5)
Ring_ShuffleShardCached/num_instances_=_1000,_num_zones_=_1,_shard_size_=_0-10         56902.00n ± ∞ ¹   69.58n ± ∞ ¹   -99.88% (p=0.008 n=5)
Ring_ShuffleShardCached/num_instances_=_1000,_num_zones_=_1,_shard_size_=_3-10            146.5n ± ∞ ¹   145.9n ± ∞ ¹         ~ (p=0.690 n=5)
Ring_ShuffleShardCached/num_instances_=_1000,_num_zones_=_1,_shard_size_=_10-10           357.9n ± ∞ ¹   358.4n ± ∞ ¹         ~ (p=0.421 n=5)
Ring_ShuffleShardCached/num_instances_=_1000,_num_zones_=_1,_shard_size_=_30-10           1.023µ ± ∞ ¹   1.021µ ± ∞ ¹         ~ (p=0.889 n=5)
Ring_ShuffleShardCached/num_instances_=_1000,_num_zones_=_1,_shard_size_=_100-10          3.604µ ± ∞ ¹   3.604µ ± ∞ ¹         ~ (p=1.000 n=5)
Ring_ShuffleShardCached/num_instances_=_1000,_num_zones_=_1,_shard_size_=_1000-10         56.92µ ± ∞ ¹   57.21µ ± ∞ ¹         ~ (p=0.841 n=5)
Ring_ShuffleShardCached/num_instances_=_1000,_num_zones_=_3,_shard_size_=_0-10         57481.00n ± ∞ ¹   69.66n ± ∞ ¹   -99.88% (p=0.008 n=5)
Ring_ShuffleShardCached/num_instances_=_1000,_num_zones_=_3,_shard_size_=_3-10            144.6n ± ∞ ¹   144.6n ± ∞ ¹         ~ (p=0.817 n=5)
Ring_ShuffleShardCached/num_instances_=_1000,_num_zones_=_3,_shard_size_=_10-10           423.3n ± ∞ ¹   417.6n ± ∞ ¹         ~ (p=0.310 n=5)
Ring_ShuffleShardCached/num_instances_=_1000,_num_zones_=_3,_shard_size_=_30-10           1.023µ ± ∞ ¹   1.032µ ± ∞ ¹         ~ (p=0.738 n=5)
Ring_ShuffleShardCached/num_instances_=_1000,_num_zones_=_3,_shard_size_=_100-10          3.747µ ± ∞ ¹   3.627µ ± ∞ ¹         ~ (p=0.151 n=5)
Ring_ShuffleShardCached/num_instances_=_1000,_num_zones_=_3,_shard_size_=_1000-10         57.83µ ± ∞ ¹   56.68µ ± ∞ ¹    -1.99% (p=0.032 n=5)
Ring_ShuffleShard_512Tokens-10                                                            173.9µ ± ∞ ¹   173.3µ ± ∞ ¹    -0.35% (p=0.032 n=5)
Ring_ShuffleShard_LargeShardSize-10                                                       8.231m ± ∞ ¹   8.217m ± ∞ ¹         ~ (p=0.095 n=5)
Ring_ShuffleShard_ShardSize_0-10                                                     5152882.00n ± ∞ ¹   61.14n ± ∞ ¹  -100.00% (p=0.008 n=5)
geomean                                                                                   26.13µ         6.128µ         -76.54%
¹ need >= 6 samples for confidence interval at level 0.95
```

**Which issue(s) this PR fixes**:

Follow up to https://github.com/grafana/dskit/pull/560.

**Checklist**
- [x] Tests updated
- [na] `CHANGELOG.md` updated - the order of entries should be `[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`