@pstibrany 2024-08-20
I've rerun `go test -run TestRing_ShuffleShardWithLookback_CorrectnessWithFuzzy -count=1000 -parallel 4 -timeout 1h` after latest changes, and it did not report any errors.

TODO: test for new `filterOutReadOnlyInstances` method.